requirements.txt
pyquaternion
tqdm
tensorboardx
pyyaml
open3d==0.11.2